Moved prometheus monitoring into flask blueprint
