+++ cellxgene_gateway/templates/cellxgene_error.html
@@ -28,11 +28,11 @@
           
      <h4>{{ message }}</h4>
 
-     <a href="{{ url_for('filecrawl') }}">
+     <a href="{{ url_for('gateway_blueprint.filecrawl') }}">
           Please click here to be redirected to the file directory.
      </a>
      <br>
-     <a href="{{ url_for('index') }}">
+     <a href="{{ url_for('gateway_blueprint.index') }}">
           Please click here to return to the homepage.
      </a>
      </div>
