+++ cellxgene_gateway/templates/process_error.html
@@ -39,10 +39,10 @@
           <li><a href="{{ relaunch_url }}">
                     Attempt to relaunch the cellxgene server.
                </a></li>
-          <li><a href="{{ url_for('filecrawl') }}">
+          <li><a href="{{ url_for('gateway_blueprint.filecrawl') }}">
                Return to the file directory.
           </a></li>
-          <li><a href="{{ url_for('index') }}">
+          <li><a href="{{ url_for('gateway_blueprint.index') }}">
                Return to the homepage.
           </a></li>
      </ul>
