+++ environment.yml
@@ -7,6 +7,7 @@ dependencies:
 - flask
 - psutil
 - black
+- typer
 - twine
 - isort
 - coverage
@@ -16,3 +17,4 @@ dependencies:
   - flask-api
   - werkzeug
   - cellxgene
+  - prometheus-flask-exporter
