+++ requirements.txt
@@ -5,3 +5,4 @@ werkzeug
 psutil
 requests
 prometheus-flask-exporter
+typer
